bumped versions for deploy test, moved tests
