+++ .github/workflows/gh-test.yml
@@ -25,6 +25,7 @@ jobs:
         python -m pip install --upgrade pip
         pip install flake8 pytest
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+        pip instal mujoco_py
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
