+++ setup.py
@@ -38,7 +38,7 @@ with open("README.md", "r", encoding="utf-8") as fh:
 
 setup(
     name='mdp-playground',
-    version='0.0.2',
+    version='0.0.3',
     author=AUTHORS,
     author_email=AUTHOR_EMAIL,
     description="A python package to design and debug RL agents",
