+++ tests/test_version.py
@@ -1,0 +1,5 @@
+from mdp_playground import __version__
+
+
+def test_version():
+    assert __version__ == '0.0.3'
